requirements.txt
discord.py
asyncio
